requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
netifaces==0.11.0
psutil==7.2.2
python-nmap==0.7.1
requests==2.32.5
scapy==2.7.0
urllib3==2.6.3
